ZK: Allow the configuration of client buffer size
The patch adds a `zookeeper.buffer_size` configuration option to
set the maximum size of the ZK client buffer. This configuration
option is currently only available as a JVM system property called
`jute.maxbuffer`.

Note that this property does not specify the size of the buffer
allocated for each read (this is equal to the number of bytes
awaiting to be read), but rather an upper bound for the maximum
buffer that can be allocated.

Ref: MNA-548

Change-Id: I03ab088259096379de74955b1bb3625ad9866074
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>