VxGW: apply backpressure on network list
This patch fixes a bug whereby the service would crash when initializing
if the virtual topology contained more than 128 networks.  The
underlying cause was a buffer size in RxJava's observeOn that is used to
deliver work units (networks) to the thread examining them and deciding
whether the network participates in a VxLAN Gateway.  This buffer has a
size of 128, and triggers backpressure on the parent observable when
full.  This would happen mainly during initialization, when all existing
networks are notified at once.  Unfortunately, the type of Observable we
were using (a PublishSubject) does not support backpressure, so this
process resulted in a failure.

This patch fixes this by applying onBackpressureBuffer on the network
observable.  This buffer is sized to a large value that can be
controlled via mn-conf, and is sufficient for our current scalability
targets.  Upon overflow, the service will alert the operator by
displaying an error message in logs.

The patch also includes two counters exposed through the cluster's JMX
bean that allow monitoring the total number of networks watched and
those that are part of a VxLAN Gateway.

Fixes:  MNA-554
Change-Id: Ib796e69d444adcf67a91d4ef1dc2e2769cb8bcfd
Signed-off-by: Galo Navarro <galo@midokura.com>