Update Pool-HM mappings on HM create/delete
Starting in Kilo, MidoNet plugin adopted the advanced service model for
LBaaS, and in this model, the mapping update of the Health Monitor and
Pool could only occurr when the Health Monitor is created or deleted.

In this patch, update the mapping correctly in the HealthMontior
Create/Delete.  Organize the code so that the legacy way (separate
mapping API calls) and the new way share as much code as possible.

Change-Id: Ib9c1cee687f1c44600b73f3423ef30fb3f205c84
Backports: MNA-552