use router interface id to get the router
The RouterInterface id is the id of the router
corresponding to the router interface. The port data
in midonet's storage may not have had the device_id
set to the router id at the time of it's creation,
and therefore we can not rely on the port data to get
the router id.

Change-Id: I9ae86e920ca10be01ef179af7df43fb60ae30e32
Signed-off-by: Joe Mills <joe@midokura.com>