FlowHistory: Catchalls for RuleResult and SimulationResult
RuleResult and SimulationResult are enums which can change between
releases. The Json flow recorder had its own representation of both of
these enums, and converts them when recording using a match. This change
adds a catchall so that if a new RuleResults or SimulationResults are
added, we do not get a match error.

Change-Id: Id2ee2f61a64bb3fcb2067b0d9f50f4f7565dc13b
Signed-off-by: Ivan Kelly <ivan@midokura.com>