Agent: Close HostRequestProxy feedback loop
A feedback loop in HostRequestProxy was causing the agent to crash:

A host with X ports bound would cause the HostRequestProxy to try and
fetch those ports, scheduling a re-sync for when we received one of
those ports that was not cached. If none were cached, then when we got
the first port we would re-sync and try to fetch X and fail for X - 1,
again scheduling a re-sync. With (X - 1) * 2 re-syncs pending, we
would continue this cycle until all ports were loaded.

Moreover, on every re-sync we would try to fetch all the ports' state
from Cassandra, in case it was down.

Change-Id: I5baab346410f5c6641ff1cc31d0911868cb1f342
Signed-off-by: Duarte Nunes <duarte@midokura.com>