Fix observeOn on topology API Service
Topology service sessions observe topology
events in their own thread via observeOn.
The internal observeOn buffer is limited to
128 elements, which may be too low for
large topologies, potentially causing a
backpressure exception.

This patch adds an onBackpressureBuffer
operator before observeOn, and a configuration
option to set the size of this buffer. If the
buffer size is exceeded, the session will be
terminated.

Ref: MNA-562

Change-Id: I93474ee820921b0f6e43136ab5bae8d5f285081e
Signed-off-by: Ernest Artiaga <ernest@midokura.com>