Agent: Install flows for expired packets
This patch fixes a regression where we stopped installing flows for
packets dropped as a consequence of being evicted from the
WaitingRoom. A packet is evicted when it times out waiting for state
to complete the simulation, like a virtual device or an ARP reply.
Like the flows we install for packets dropped because of errors
during the simulation, this short lived flow signals that currently
we can't process the matching packets, and applies backpressure to
the sender or senders.

Change-Id: If65bd6f8f778e3200de7e747ae89e4e4410979a2
Signed-off-by: Duarte Nunes <duarte@midokura.com>