L2-based create fake uplink scripts
The new architecture of Midonet allows for L2 to be the uplink network
of the cloud, instead of requiring the Provider Router as before, which
is the most common set up of Neutron.  These scripts set up and destroy
the local host environment to be able to access the uplink ('external'
in Neutron) network from the host.

Change-Id: I5e0910eec7609e8518b925ce7d9395b86634f816
Implements: MNA-563