Update patch set 4
Patch Set 4:

We did manual testing with wireshark. I could add an integration test that makes sure that, for equivalent packets, the checksum we produce is the same as the checksum that, for example, linux generates. I'd do that for UDP too, ICMP and all combinations of packets. I'll create an epic :p

Patch-set: 4