Update ARP entry for FIP IP & Router gatway MAC.
Add an ARP entry for the Floating IP's IP address and the associated
tenant Router's gateway port MAC address to the extenal Network's ARP
table upon FIP creation or update. Delete the ARP entry upon deletion of
Floating IP or update that moves the Floating IP to a different router.
Finally remove the logic to update gateway routes with the router.

Change-Id: I1fd77157e78d2731eb67537eb9bc418267d2ac1d
Ref: MNA-460