mn-conf(1) man page: Fix angle bracket in set command
The ronn tool gets confused when a single opening angle bracket is
contained within backticks (as part of a command), followed by further
angle brackets (for a variable).

This results in two angle brackets being displayed and the following
variable not being marked as such.

This patch escapes the angle bracket character and makes it bold.

Additionally, the description of the command has been changed to make
it clearer that the settings are read from a file.

Fixes: MNA-559

Change-Id: Ia6d514b6912a292673c118304ac77d31ee6d9990
Signed-off-by: Jan Hilberath <jan@midokura.com>