Update patch set 4
Patch Set 4:

Okay, but then we'll have more of those tests with bytes hardcoded, which is basically an unmaintainable mess.

As for being TCP only, I was saying that none of the other checksum functions have unit tests. (This is basically the one we use for UDP.) If we are worried about changing the checksum functions such that we need unit tests to catch breaking changes, we should unit test all of them.

Patch-set: 4