ZOOM: Remove object ownership
This patch removes the object ownership from ZOOM, which was
previously used to save state information. With the introduction of
the state storage, this feature is no longer needed. By removing it,
the patch also improves the storage performance by halving the
number of reads for each write operation (ZOOM no longer reads both
the object and its owners).

Change-Id: I210d9f104311ba4a362f6ae618efd72d8f5c65e3
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>