Update patch set 4
Patch Set 4:

It's true, the pre-existing functions don't have unit tests. But still, I think newly added code should come with tests. The packets library has had bugs in the past, we should learn from that mistake.

These byte blobs don't have to be maintained, as the TCP and IP protocols are not going to evolve on us, so that's not a worry.

Patch-set: 4
Label: Code-Review=0