ZOOM: Error handling for object observables
This patch replaces the InstanceSubscriptionCache with the
NodeObservable for ZOOM object observables. The NodeObservable
clears any watcher in ZK when the last subscriber unsubscribes,
and closes the observable with a specific error.

The new implementation will automatically recover any new subscriber
on a closed observable, such that it is not possible to subscribe
to an observable that no longer has an active watcher in ZK.

This mechanism also ensure that there is no need for a custom
mechanism to remove closed observables from the ZOOM observables
map, since when any subsequent subscription will create a fresh
observable and add it to the observables map.

Ref: MNA-129

Change-Id: I8a000a2dea4779e1e33667fded5a1c8e9c7b2338
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>