Update patch set 2
Patch Set 2:

Alex, thanks for the review. I've fixed all comments, except for the replacements of throw exceptions in favour of Responses.

This is because the Response needs to get entities, which in turn need providing error serializations. (e.g, we want to feed the violations into the response). This works nicely via exceptions, so I preferred to keep it like that for now.

Patch-set: 2
Label: Code-Review=0