FlowTracing: only add trace request id to context once
Fixes the issue where a single trace message is being output more than
once. The problem was that the trace request id was being added to the
list of trace request ids more than once. This was a set previously, so
adding again wasn't an issue, but now as it is a list, previous
existence must be checked.

Ref MN-3491

Change-Id: I8dadd93714a90a10dfa276b38030464dc3e4937e
Signed-off-by: Ivan Kelly <ivan@midokura.com>