DRAFT - Cluster API: introduce static file server
This patch introduces a new path in the API endpoint, that points at a
configurable local path on the host and will serve any files under it.
This configuration is made via mn-conf, using two new properties added
to the cluster.rest_api minion:

    cluster.rest_api.static_path=/usr/share/midonet-api/static
    cluster.rest_api.static_uri=/static

With the default values provided, if we place a file named "sample" in
/usr/share/midonet-api/static/sample, then a GET on this
http://<ip>:<port>/static/sample will serve the file, inferring the
mime type from the file.

This feature can be used to deploy, for example, configuration documents
that can be at a convenient location for operators.

Change-Id: I561b56608624fc03c83f6abaddd883203cd8d408
Signed-off-by: Galo Navarro <galo@midokura.com>