Cluster: support embedded ZK
This patch provides support for running a ZK cluster embedded in the
MidoNet Cluster.  A cluster node will start up, and based on its local
configuration file start a local ZK node that will try to join an
ensemble described by the zookeeper.zookeeper_hosts property in config.

Until now, the Cluster expected a ZK to exist, and the Configuration
system would try to load config from there before starting up.  This is
no longer true, so we needed to tweak how we load and pass config to the
various Guice services that are brought up in the cluster.  This leaked
slightly into midolman due to interface changes in MidonetBackendModule.

Change-Id: I05b20bab0b21c312feb0fcd8fcea9491b88a17bb
Signed-off-by: Galo Navarro <galo@midokura.com>