API: Upgrade Jackson to 2.5.4
This patch upgrades the Jackson version to 2.5.4, as follows:
- version for MidoNet from 1.9.3
- version for ODL-OVSDB fork from 2.3.2

From the perspective of the MidoNet codebase, the most signifcant
change due to the upgrade, is that Jackson 2.5.4 includes
cached end-point readers and writes. Therefore, it is possible
that during JSON serialization/deserialization the MidoNet
implementation of the ObjectMapper will not be called, but rather
the cached version of a reader/writer used during a previous
serialiaztion operation.

Therefore, to accommodate the version upgrade, the patch also
includes the following changes.

- A `MidonetObjectReader` implementation of the Jackson
`ObjectReader` to handle JSON serialization exception thrown
during deserialization (previously all these exceptions were
handled by the MidonetObjectMapper).

- The JAXB JSON provider intercepts all serialization/
deserialization operations to add custom `Version` annotation
to the key indexing the cached readers and writers. This
ensures that given and end-point class, the appropriate
reader/writer is selected, according to the version specified
in the request media type.

Change-Id: Ic123547c1faf7dc6a70ed3d4ebec36fa8e999df0
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>