Compat API: Use legacy API JAXB JSON provider
This changes the compatibility API to use the same JAXB JSON
provider used by the legacy API.

The initial implementation of the compatibility API introduced a
barebone provider, designed to accommodate most serialization
scenarios. This was because at the time the legacy API provider
was implemented as part of the `midonet-api` module. However, this
provider does not provide full support for all API models, which
includes specific serialization options, API versioning, etc.

With the migration of the legacy JAXB JSON provider to the `cluster`
module, the compatibility API can levereage it for JSON
serialization.

Ref: MNA-584

Change-Id: I3db079b21e60a9f055d5de534fd09f055a10895e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>