Agent: Cache Workers in UpcallConnectionManager
When binding a lot of ports (upward of 2k), the agent was exhibiting
timeouts when asking the PacketsEntryPoint for the Workers object.
This is due to the fact that the same thread is used for the
DatapathController and the PacketsEntryPoint, which was busy with
binding all of these ports.

This patch caches the response so we only request it once.

Ref: MN-4100
Change-Id: I1cc18b2f276a4355558aa6809427a75d02f994b8
Signed-off-by: Duarte Nunes <duarte@midokura.com>