Make NeutronTranslator persist Neutron model.
Extend NeutronTranslator base class with retainNeutronModel so that it
would also handle the persistence of the original Neutron model, which
used to be handled directly by C3POStorageManager. The method may be
overridden so that each concrete Translator can choose whether or not
and how to persist the Neutron Model. (E.g. RouterInterfaceTranslator)
This allows us to manage back-references between Neutron models by
properly populating, copying and deleting back-references.

Change-Id: I26537872d6b47827061dc79b832c6bc3cf5e33ed
Ref: MNA-281