Add backreference to Floating IP to Neutron Port.
Add backreference to Floating IP to Neutron Port model so that it could
allow proper cleanup such as removal of an ARP table entry for Floating
IP upon deletion of the port to which the Floating IP is attached.
Since Neutron does not manage such a backreference, PortTranslator needs
to use NeutronPortUpdateValidator to restore the backreference upon
NeutronPort update.

Ref. MNA-281

Change-Id: I53a1c1562288e2f9671b02da3eed01b373bc2cfb