Update patch set 1
Patch Set 1:

Thank you for your feedback, Brandon!

I'll prepare a patch shortly that'd do what I really need to do. The main issue is that the notion of Floating IP is a Neutron Concept, and in MidoNet, it's just a pair of NAT rules. We need to delete the ARP entry for Floating IP upon deletion of the Port (VIF port, etc) which the Floating IP is bound, but Neutron doesn't send a DELETE op for Floating IP. Therefore we need to look up a Floating IP from a normal Neutron Port.

One option is, as this patch does, to add a reference to the Floating IP from the Port (that is, its fixed IP). We can alternatively make MidoNet Port to refer to Neutron Floating IP or maybe just make the port has an extra IPAddress field containing the Floating IP address. But either way, I don't like mixing in Neutron Concepts to MidoNet models.

What do you think?

 > (1 comment)
 > 
 > I'm not sure I understand what this is for. Right now you're just
 > adding and maintaining the reference; maybe this patch should
 > include the part where you actually use it?
 > 
 > Also, whatever problem this is intended to solve, can it be solved
 > by adding more information to the Midonet topology objects?
 > 
 > There's been some talk about moving away from relying on the
 > Neutron objects in Zookeeper during translation, and instead
 > storing all the information we need in the Midonet topology
 > objects. If it's not much harder, I think we should do that in this
 > case.

Patch-set: 1
Label: Code-Review=0