Catch StateAccessException in HealthMonitor
Due to a topology bug, it is possible that a mapping info object
for health monitor may exist, but the pool or load balancer object
may not. This situation can be caused by a different bug, but needs
extra care for handling here to prevent the agent from continuously
restarting. Because we don't catch this exception, it is thrown up
to the Actor System level, which will then restart the whole of
midolman.

Change-Id: I9a83c7babede4bca35fc4750f07b9bd519b1981f
Signed-off-by: Joe Mills <joe@midokura.com>