Agent: Clear the PacketContext when dropping a packet
When we are dropping a packet as a result of an exception (or from any
other code path that's not a packet expiration in the WaitingRoom), we
are not clearing the PacketContext, namely flow tags or actions. This
may cause extra memory to be used (for holding flow tags), or worse,
actions to be installed. This patches fixes these problems.

Change-Id: I1590a05ab93ecb72a2ca85bcebf7f8535b797408
Signed-off-by: Duarte Nunes <duarte@midokura.com>