Add Mirror rule
This patch adds a new rule for port mirroring. Users can create a mirror
rule with a destination port ID and add it to a chain although this
patch doesn't add the actual port mirroring capability in Midolman.

The actual simulation will be added by another patch.

This resolves MNA-240 and MNA-605.

Change-Id: I2790e7583c30fa9dbc13ff4bc00aa94b22b23950
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>