Agent: Don't create trace state for egress packets
This patch ensures we don't create trace state for egress packets,
such as when we receive a tunneled packet that we trace. Right now, we
create trace state for these packets, which causes us to get into the
FlowStateReplicator and replicate that state, ultimately resulting in
a NullPointerException because no ingress port is set.

Ref MN-3658

Change-Id: Idbd467f81f3d2a7994b77aa77395a46bcbd2ac65
Signed-off-by: Duarte Nunes <duarte@midokura.com>