Add device type information for device tags
For flow history it is interesting to know the type of device traversed
as well as the device id. This could be looked up from topology, but
doing so would incur many reads (the id would have to be probed against
each device type). This information is available when tags are added, so
it's simple to just attach it to the tag. This patch does that.

Ref MN-3608

Change-Id: Iad3075755c122a2619b2c9d88a025b612c5d2764
Signed-off-by: Ivan Kelly <ivan@midokura.com>