Moving topology setup to per-binding instead of per-suite
Test test_router:test_spoofed_arp_reply was not correct for the second
binding as it was reusing the virtual topology of the first binding
(the router used on the second binding already learned the spoofed
mac address). This fix addresses this and improves the isolation
between bindings by creating and removing the whole virtual topology
for each binding.

Pros: better isolation among test bindings.
Cons: slower runs as the virtual topology needs to be re-created for
each binding.

Change-Id: Ifa483fda478cf44e4a2f2158d5814903b934cb2f