Agent: Remove warning about dropped packets not being idle
When dropping a packet because it was expired or because of an error,
we are not removing them from the waiting room. Instead of ensuring we
do so, this patches simply ignores such packets when evicting them
from the WaitingRoom, as this plays well with future changes where the
WaitingRoom becomes a ring buffer contained pre-allocated
PacketContext's.

Change-Id: Ice2c0ee0cacaa6c80a194b7b3c2e9b43eca36271
Signed-off-by: Duarte Nunes <duarte@midokura.com>