A service discovery framework
This patch introduces a very basic framework for service discovery. It
allows a service to register its availability to offer a given
service, and a client to obtain the list of registered services. It's
up to the client to decide which one to pick at any given time.
More sophisticated policies are left for following patches. Also added a
mock discovery service for easier testing in other modules.

Ref: ME-431 MI-653

Change-Id: I024cc9a9f6f102d77a0b20beabb6f33aee5381da
Signed-off-by: Ernest Artiaga <ernest@midokura.com>
Signed-off-by: Xavi León <xavi.leon@midokura.com>