Update patch set 8
Patch Set 8:

gate_midonet-api is failing because of a new validation utilises MidonetBackend which is not in the Guice dependency in the old REST API. Currently org.midonet.cluster.rest_api.models.Rule is shared among the new cluster REST API and the old REST API. So we need to switch to the new API or to keep 1.x code and test against the old REST API for now. I'll leave 1.x code until this issues is resolved and move on toward the agent simulation code because of the soft deadline if the task. I need to switch to another task in short.

Patch-set: 8
Reviewer: Taku Fukushima <1002743@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0