Support new BGP model in the new architecture
This patch modifies the Protocol Buffer model definitions, the
compatibility API and the BGP mapper to use the new BGP model.
In this model, all BGP information (peers are advertised networks)
are stored in the router port. Furthermore, the BGP mapper exposes
the BGP updates as a BgpRouter instance, which contains all
peers and networks configured at a time.

Ref: MNA-466

Change-Id: I618a246cc56b272ab982750a4a73c467b399d8fb
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>