Add Rule<->Chain Zoom bindings
We were managing these explicitly before, and not always
correctly. With Zoom bindings, it's now possible to add a Rule to the
end of a Chain just by creating the Rule and specifying the
chain_id. To add the Rule to another position in the Chain, we still
need to create the Rule first (no need to set the Rule's chain_id) and
then create/update the Chain with the Rule's id inserted into its
rule_ids list in the desired position.

Also fixes some issues with the V2 REST API's handling of Rule
positions, which fixes some failing unit tests.

Change-Id: I701370cf4723d5910a4912eba32f7d73cd54c001
Signed-off-by: Brandon Berg <bberg@midokura.com>