Agent: Conservatively mask IEEE 802.3 packets
Dropped IEEE 802.3 packets, for example because they ingressed a
router port, would lead to a flow mask that only matches the input
port number, thus causing all incoming traffic on that port to be
dropped for the duration of the drop flow.

This patch adds a branch to the FlowMask code to just create a
specific flow for these types of packets.

Change-Id: I6606fef0ef894429fd3b6c214a32a15d1a0f5204
Signed-off-by: Duarte Nunes <duarte@midokura.com>