API: Exclude non relevant tests from new stack
This patch disables a bunch of TestVtep tests and all TestTraceRequest
when running against the Cluster REST API.  This will be here until we
migrate these features over to the new stack.

Also making pass a couple of TestVtep tests that were failing due to
ApiException not initializing the Message in the super class.

Change-Id: I1d3c213ba6b187abd95467ecc5e3f06275dcfbbb
Signed-off-by: Galo Navarro <galo@midokura.com>