Update patch set 4
Patch Set 4:

What do you guys think about our MD proxy calling HTTPServletRequest#getRemotePort to get the source port of the request, then based on that decide what network the request came from?

I was looking for a way to signal to our MD proxy what the source network/VM is, so that it doesn't have to listen on one port per network (or worse, spin up one proxy per network). This was the best I could think of.

MidoNet could translate the source port of the original request into a specific range, depending on what network the VM is in - and of course reverse translate when the proxy sends response packets back.

Patch-set: 4