Agent: Clear trace state transaction
Currently, the flow trace transaction is not cleared when the
simulation completes, so that it can persist across a
TraceRequiredException. This means it won't be clears when a
simulation that doesn't commit state transactions completes, like is
the case of ErrorDrops for example. It is also not cleared when an
error is thrown. It is only cleared when a packet is restarted with
a NotYetException.

This patch fixes this by ensuring the trace state is correctly
cleared.

Ref MN-3703

Change-Id: Ib1244999fea11b6a3a0fbe99215d773f67d4638f
Signed-off-by: Duarte Nunes <duarte@midokura.com>