Specify router GW port for FIP NAT rule i/o port.
Specify Mido Router Gateway Port to in/out port of the D/SNAT rules for
the Floating IP. Currently Floating IP Translator passes Neutron Router
Gateway Port, which is actually a Neutron Port on external Network
linked to the Rotuer Gateway Port. We need to pass the Router Port so
that packet simulation would be able to correctly match the in/out port.

Fix MN-650

Change-Id: Ibe960fe016d9bc44714996174b1489cd38f7371f