Update patch set 2
Patch Set 2:

Pino, in Takashi's current proposal, NAT rules (to link local address) are used to distinguish the VMs on the same host.  Also, these flows are inserted pro-actively since metadata traffic don't really require user-space intervention.  Your suggestion certainly works too, but just want to know why you think your way is better than the one proposed.  Is it because it's easier to implement?  I was thinking having a 'special handling' in the agent for this traffic is better avoided, but since I'm not familiar with the agent code, I will leave it to the agent team to decide.

Patch-set: 2