Update patch set 7
Patch Set 7: Code-Review-1

(15 comments)

A few comments mostly regarding model discrepancies.

Also: Maybe there's no other way, but I was expecting the mapping to be super compact, straightforward code, as it just needs to produce full objects out of Zoom, all of the diffing code is written in the RoutingHandler already.

Finally. The RoutingHandler needs a small update:

Currently it receives the full router bgp config object but in v1 this is a trick: it's only filled in with what's relevant for the port it is handling.

In v2 it should filter out peers not reachable though its port and just ignore them, because it really will be getting the full bgp config for the router.

Patch-set: 7
Label: Code-Review=-1