Update patch set 9
Patch Set 9:

(1 comment)

The update LGTM, we should discuss this comment though:

"Finally. The RoutingHandler needs a small update:
Currently it receives the full router bgp config object but in v1 this is a trick: it's only filled in with what's relevant for the port it is handling.
In v2 it should filter out peers not reachable though its port and just ignore them, because it really will be getting the full bgp config for the router."

Patch-set: 9
Label: Code-Review=0