Update patch set 1
Patch Set 1:

Brandon, right now Neutron always deletes the DHCP before its ports, so no cleanup is in reality necessary, but I don't know if that's an assumption we can always make. And if a DHCP is not found in non-delete cases, then that's an invalid state and should throw an exception.

Patch-set: 1