Condition EtherType handled as positive integers
This patch fixes an issue where the EtherType field for rule
conditions is stored in ZooKeeper as negative numbers. This happens
for rules created from Neutron security groups, with EtherType
values that exceed the value 32767 (0x7FFF). This happens because
the EtherType values are handled internally as short 16-bit values, yet
they are written to storage as 32-bit integers without removing the
negative sign.

These changes ensure that the proper sign conversion is performed
before an EtherType larger that 0x7FFF is written to storage.

It also ensures backward compatibility with topology data from
existing deployments, by performing the conversion to a positive
EtherType value for conditions that have been created before this
patch. This guarantees that conditions are properly matched
for existing traffic.

Fixes: MN-3705

Change-Id: I4eb9be61f1f8cdf4651e1282bf6be434a899c220
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>