Fix in-memory storage notifications
This fixes an issue in the in-memory storage, ocurring when an
observer subscribes, while being notified of an update during a
transaction, to an instance observable for an object modified during
the same transaction. In such circumstances, it is possible that the
observer will receive a stale update for the object, rather than
the value being given to it by the transaction.

Change-Id: I9b9fc8c315e58573ed0ce3395337848199283f63
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>