MacTable based on a Merged Map
This patch provides an implementation of a mac table based on a merged
map. A flag called use_merged_maps in KafkaConfig determines whether
merged maps are enabled. When this flag is false we fallback to
replicated maps.

In order to do so, we modify the ZookeeperLegacyStorage such that
it not only implements trait LegacyStorage but also StateTableStorage.
As we migrate away from replicated maps, methods to obtain state
tables in trait LegacyStorage will migrate to trait
StateTableStorage.

ZookeeperLegacyStorage now also takes a merged map bus builder as
an argument to enable the creation of merged maps. The builder
comes in two forms: one for real deployments and one for testing.
In the latter case, the merged map bus is in-memory.

To minimize changes to the legacy code (in particular to the
managers), the legacy MacLearningTable trait extends
StateTable. This way, the BridgeManager can build a simulation
bridge, which now takes a state table, with a MacLearningTable,
and the BridgeMapper can build a bridge with a state table.

Ref: MNA-482

Change-Id: I4fd67a0904848073035ffb879fc308560a144ee6
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>