MacTable based on a Merged Map
This patch provides an implementation of a mac table based on a merged
map. The mac table of the bridge mapper relies on a merged map when
merged maps are enabled. In production property kafka.use_merged_maps
must be set to true in MidoNet's configuration file. For tests merged
maps are enabled by setting system property kafka.use_merged_maps to
true.

Ref: MNA-482

Change-Id: I4fd67a0904848073035ffb879fc308560a144ee6
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>