Clean up translator IT task creation
Make C3POMinionTestBase.insertTaskSql and .executeSqlStmts private,
and replace all uses in derived classes with insertCreateTask,
insertUpdateTask, or insertDeleteTask. Also make the these functions
take a JsonNode instead of String as the json argument, eliminating
the need to call .toString() on the json node every time these
functions are called.

Change-Id: I46b3673d5881ce51e69292c8c459f0d15b8f0511
Signed-off-by: Brandon Berg <bberg@midokura.com>