Make VirtualTopology tests single threaded
VirtualTopology and the in-memory storage class it uses currently create
their own threads and schedule all observable work on these
threads. This makes it hard to assert in tests whether all notifications
that will happen for an update will happen or not. It also makes the
tests non-deterministic, as changes in thread scheduling could change
the order of notifications with regard to the threads which are checking
for them.

This patch makes 2 major changes. First is that VirtualTopology now gets
its executors passed in through guice. This allows the unit tests to
pass in a different executor to the one used by integration testing and
production. This executor should behave similar to a single thread
executor, so I've also added a new executor type, SameThreadButAfter,
which will run any executions in the same thread as the one submitting,
but if executor is already running something, the new runnable will be
added to the end of a queue, and will only be run after the current
execution finished.

The other major change is to remove threading entirely from
InMemoryStorage, as this is only ever used in unit tests.

With these changes, only one test in ChainMapperTest failed, which
proved to be a problem with the test, so that has also been updated.

Change-Id: If84eaf12327af47d268cdf6a8c9f3708c4fc0d5c
Signed-off-by: Ivan Kelly <ivan@midokura.com>