Fix Flakiness in MergedMapTest
This patch fixes flakiness in MergedMapTest:
-It breaks down the invalid opinion scenario into two. This is needed
because invalid opinions complete the merged map's observable.
-Scenario "Remove non-existing opinion" is also fixed. Previously this
test would result in a Timeout Exception when the just inserted
opinion was removed before the observer gets notified. As a result no
notification would be sent to the observer instead of two. We now wait
to get notified for the opinion insertion before removing it.

Change-Id: I2ed4f029c8e429ebacd300cff023d6b54f2341da
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>