[WIP] Add port mirroring to the simulation.
Previously we squashed rule applications into a PacketContext instance
and generated flow actions in the end of the simulation. With this patch
a new mirror rule takes a snapshot of the current state of the
PacketContext and generates interim flow actions at the point where the
mirror rule is applied. In other word, the simulation is concluded
everytime when the application of the mirror rules happen and the
simulation keeps going until it reaches the bottom of a chain.

In the current implementation the simulation translates every mirror
rule even if they're duplicated. So users MUST make it sure that the
port mirroring is correctly done.

Change-Id: I34ecf6d8ff411897cbbbc538589690286abfe551
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>