MergedMap Notifications and close()
This patch changes the notification type of merged maps from (key,
value) to (key, oldVal, newVal). This is to make merged maps better
fitted to implement mac and arp tables. Additionally, this patch
provides a way to "close" a map. This completes the map's
observable and clears its locally cached content.

Change-Id: I0141833dc4e468f825fbc27fe4f8b0feef5dee73
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>