ClusterBridgeManager: fix zk watcher leak
The watcher for the list of bridge logical ports was fetching
the list of exterior bridge ports and setting itself as a watcher
from it. Because this watcher itself is leaky, in that it re-fetches
everything and reinstalls itsef as watcher of several objects, adding
a feedback loop triggered by changes in the list of exterior ports was
causing an exponential increase in watcher executions, these would
lead to thousands of redundant invocations that would induce lags of
minutes in the ZK reactor thread.

Fixes: MN-3704

Change-Id: Ibc47adb0fced52bb1a3440c20424969ea9c1f4ca
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>