Agent: Garbage collect unused NAT blocks
Periodically run a task that clears unused NAT blocks.We reclaim
blocks at the IP granularity, meaning that for blocks to be
reclaimed all blocks for a particular IP must be unused.

We protect against a concurrent process claiming a block and also
against a concurrent GC deleting these blocks.

Ref MN-3723

Change-Id: I013ed7dfcde908aaa19ad70c02ab34db59224421
Signed-off-by: Duarte Nunes <duarte@midokura.com>