Update patch set 2
Patch Set 2:

(1 comment)

We need to be able to test both the old stuff and the new stuff until we're ready to rip out the old stuff when the new stuff is 100% ready. For this we have the VirtualConfigurationBuilders, which is a facade in front of the storage layer allowing you to build a virtual topology without caring about the backend. LegacyVirtualConfigurationBuilders is for the old stuff. ZoomVirtualConfigurationBuilders is for the new stuff. There's a switch, -Pmidonet.newStack for switching between them.

Patch-set: 2
Label: Code-Review=0