PortTranslator.translateUpdate() preserves binding
PortTranslator.translateUpdate() was deleting port binding information
due to an incorrect understanding of the NeutronPort data model,
causing a port to be permanently disabled after setting adminStateUp
to false, even after setting it back to true. For VIF and DHCP ports,
binding information is managed by mm-ctl and should not be modified by
PortTranslator.

Change-Id: Id871724262f786f8f021841867a7bf887c700200
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MNA-633