VxGW: reset watcher after update
The VxGW service was not resetting a watcher on the network when this
was notified.  This caused a bug when a network was unbound from the
VTEP, then rebound again, as the lost watcher was the one responsible
for detecting the new binding and bootstrap the VxGW Manager for that
bridge.

Change-Id: I456413aaa9a88a3dbc9e143a2a3dc8f30c6d0bd2
Fixes: MN-3745
Signed-off-by: Galo Navarro <galo@midokura.com>